io: cdedb: Fix index for multi track events
Currently, there is code, that expect, that the index field of the
courses goes from 0 to courses.len()-1 without any holes. However,
the index is filled from the index of the course in the courses
list. When there is more than one track, numbers are omitted, so that
the maximum of the index field is higher. To compensate for this, a
additional index is introduced, that is only increased when a course
for the current track is added.