Fix calculation of number of dummy participants when non-instructors without choices are present
This may happen when the simple JSON format is used: It allows to
created particpants without chaoices, which are not an instructor
either. Since these participants are always skipped in the assignment,
but have not been counted as course instructors, we have created too
few dummy participants. This resulted in an out-of-bounds error when
constructing the skip_x array.